feat: Add colored (pretty) console logging functionality
A new file `prettyConsoleTrasnport.transport.ts` was added to provide better console logging with the help of 'chalk' and 'util' npm packages. Those packages have also been added to the dependencies list in `package.json`. A 'PrettyConsoleTransport' was also exported from the `transports` folder to make the new feature available for use.